Unify tests to DRY
